move msg content util to separate file
